? need to use reference otherwise I remove a copy, not the intended unit
wip: stopping point.

need a bool to control flow, let's try this.

this time everyone got the message.  However, went to enemy natural and stopped.  So my current solution isn't optimal and is prone to failure.

I think when OnUnitEnterVision adds to buildings_enemy, is a cause of duplicates.  I have no check if stored, I just add in.

just when I thought I had it with a simple line of code ... LOL
I seem to be looping more duplicates.

missing parantheses?

comments